render: draw coordinates are f64

+++ src/render/opengl.rs
@@ -313,8 +313,8 @@ impl RendererTrait for RendererImpl {
     fn draw_sprite(
         &mut self,
         texture: &AtlasRef,
-        x: i32,
-        y: i32,
+        x: f64,
+        y: f64,
         xscale: f64,
         yscale: f64,
         angle: f64,
